Refactor changelog rendering in build script
A small refactoring was performed in the build.gradle.kts file. No functional changes were made. The change mainly simplifies the rendering of the changelog items - specifically regarding the retrieval of the change notes and how they are configured for presentation. This conforming is meant to increase readability and maintainability of the code.